requirements.txt
colorama==0.4.6
customtkinter==5.1.2
darkdetect==0.8.0
Pillow==9.4.0
pypng==0.20220715.0
pywin32==305
qrcode==7.4.2
typing_extensions==4.5.0
